cosmetic changes

+++ MSAddition/NSDate+MSAddition.h
@@ -10,31 +10,30 @@
 
 @interface NSDate (MSAddition)
 
-+(NSDate*)ms_dateWithYear:(unsigned)year month:(unsigned)month day:(unsigned)day
-                  hour:(unsigned)hour minute:(unsigned)minute second:(unsigned)second;
++(NSDate*)ms_dateWithYear:(NSUInteger)year month:(NSUInteger)month day:(NSUInteger)day
+                     hour:(NSUInteger)hour minute:(NSUInteger)minute second:(NSUInteger)second;
 
-+(NSDate*)ms_dateWithYear:(unsigned)year month:(unsigned)month day:(unsigned)day
-                  hour:(unsigned)hour minute:(unsigned)minute;
++(NSDate*)ms_dateWithYear:(NSUInteger)year month:(NSUInteger)month day:(NSUInteger)day
+                     hour:(NSUInteger)hour minute:(NSUInteger)minute;
 
-+(NSDate*)ms_dateWithYear:(unsigned)year month:(unsigned)month day:(unsigned)day
-                  hour:(unsigned)hour;
++(NSDate*)ms_dateWithYear:(NSUInteger)year month:(NSUInteger)month day:(NSUInteger)day
+                     hour:(NSUInteger)hour;
 
-+(NSDate*)ms_dateWithYear:(unsigned)year month:(unsigned)month day:(unsigned)day;
++(NSDate*)ms_dateWithYear:(NSUInteger)year month:(NSUInteger)month day:(NSUInteger)day;
 
-
--(unsigned)ms_day;
--(unsigned)ms_month;
--(unsigned)ms_year;
--(unsigned)ms_minute;
--(unsigned)ms_hour;
--(unsigned)ms_second;
--(unsigned)ms_weekday;
+-(NSUInteger)ms_day;
+-(NSUInteger)ms_month;
+-(NSUInteger)ms_year;
+-(NSUInteger)ms_minute;
+-(NSUInteger)ms_hour;
+-(NSUInteger)ms_second;
+-(NSUInteger)ms_weekday;
 
 -(BOOL)ms_isEqualDay:(NSDate*)date;
 -(BOOL)ms_isEarlierThan:(NSDate*)date;
 -(BOOL)ms_isOlderThan:(NSDate*)date;
 
-+(unsigned)ms_lastDayOfMonth:(int)month;
-+(BOOL)ms_isYearLeap:(int)year;
++(NSUInteger)ms_lastDayOfMonth:(NSUInteger)month;
++(BOOL)ms_isYearLeap:(NSInteger)year;
 
 @end
